requirements.txt
requests  
numpy
pyaudio~=0.2.11
websockets~=10.2
ws4py~=0.5.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6 
flask~=3.0.0
openpyxl~=3.0.9
flask_cors~=3.0.10
websocket-client
azure-cognitiveservices-speech
aliyun-python-sdk-core
simhash
pytz
gevent~=22.10.1
edge_tts~=6.1.3
pydub
langchain==0.0.336
chromadb
tenacity==8.2.3
pygame
scipy
flask-httpauth